refactor: 포트폴리오 페이지 폴더구조 변경
